Refactor Discord bot
